// naherb-web/src/services/generated/model/paymentMethod.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * NaHerbs Website MVP API
 * API contract for naherb-api (Spring Boot) used by naherb-web (Next.js). Updated for UUID IDs across accounts/profile/address/order/cart and Vietnam two-level address fields: province/city and ward/commune only.
 * OpenAPI spec version: 2.2.0
 */

export type PaymentMethod = typeof PaymentMethod[keyof typeof PaymentMethod];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const PaymentMethod = {
  COD: 'COD',
  BANK_QR: 'BANK_QR',
} as const;